Charset tests module
Created a new test module for charset tests
