+++ src/tests/mod.rs
@@ -98,5 +98,6 @@ macro_rules! mk_tests_default {
     };
 }
 
+mod charset;
 mod params;
 mod row;
