+++ src/tests/row.rs
@@ -12,81 +12,6 @@ mk_tests_default! {
     use rand::{distributions::{Alphanumeric, Standard}, Rng};
 
     #[test]
-    fn column_with_charset_none() -> Result<(), FbError> {
-        use crate::charset::ISO_8859_1;
-
-        // This test reproduce the of using a diferent
-        // charset of column in insert.
-        // When we read an ISO8859_1(or others) data using
-        // the UTF8, we will get an "invalid utf-8" error.
-        // In this cases, we must use the same charset
-        // of data inserted
-
-        let mut conn = cbuilder().connect()?; // utf-8, but is not used
-
-        conn.execute("DROP TABLE RCHARSETS", ()).ok();
-        conn.execute(
-            "CREATE TABLE RCHARSETS (a Varchar(30) CHARACTER SET none)",
-            (),
-        )?;
-
-        conn.execute("insert into RCHARSETS (a) values (cast('pão de queijo' as Varchar(30) CHARACTER SET ISO8859_1))", ())?;
-
-        let err: Result<Option<(String,)>, FbError> = conn.query_first("select * from RCHARSETS", ());
-        assert!(err.is_err());
-        assert_eq!("error: Found column with an invalid UTF-8 string: invalid utf-8 sequence of 1 bytes from index 1", err.err().unwrap().to_string());
-
-        // Hmm, I need use the same charset of inserted content
-        let mut conn = cbuilder().charset(ISO_8859_1)
-            .connect()?;
-
-        let (pao,): (String,) = conn.query_first("select * from RCHARSETS", ())?
-            .unwrap();
-
-        assert_eq!("pão de queijo", pao);
-
-        Ok(())
-    }
-
-    #[test]
-    fn stmt_charset() -> Result<(), FbError> {
-        use crate::charset::ISO_8859_1;
-
-        let mut conn = cbuilder().charset(ISO_8859_1)
-            .connect()?;
-
-        let (pao,): (String,) = conn.query_first("SELECT cast('pão de queijo' as Varchar(30)) FROM RDB$DATABASE;", ())?.unwrap();
-        assert_eq!("pão de queijo", pao);
-
-        Ok(())
-    }
-
-    #[test]
-    fn cast_charsets() -> Result<(), FbError> {
-        let mut conn = cbuilder().connect()?;
-
-        let qsql = "select
-                            CAST('olá abc ë' AS VARCHAR(10) CHARACTER SET WIN_1252),
-                            CAST('olá abc ë' AS VARCHAR(10) CHARACTER SET ISO8859_1),
-                            CAST('olá abc ë' AS VARCHAR(10) CHARACTER SET utf8),
-                            CAST('olá abc ë' AS VARCHAR(11) CHARACTER SET none)
-                       from RDB$DATABASE;";
-
-        let (win, iso, utf, none,): (String, String, String, String,) = conn.query_first(qsql, ())?
-            .unwrap();
-
-        assert_eq!("olá abc ë", win);
-        assert_eq!("olá abc ë", iso);
-        assert_eq!("olá abc ë", utf);
-        assert_eq!("olá abc ë", none);
-        assert_eq!(win, iso);
-        assert_eq!(win, utf);
-        assert_eq!(win, none);
-
-        Ok(())
-    }
-
-    #[test]
     fn boolean() -> Result<(), FbError> {
         let mut conn = cbuilder().connect()?;
